fix(platform): organize presenter directories
