+++ apps/platform/src/lib/infrastructure/common/presenters/get-home-page-presenter.ts
